--- packages/twenty-server/src/engine/utils/serverUrl.ts
@@ -1,18 +1,0 @@
-const ServerUrl = (() => {
-  let serverUrl = '';
-
-  return {
-    get: () => {
-      if (serverUrl === '') {
-        throw new Error('ServerUrl is not initialized. Call set() first.');
-      }
-
-      return serverUrl;
-    },
-    set: (url: string) => {
-      serverUrl = url;
-    },
-  };
-})();
-
-export default ServerUrl;
